# backend/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plan2Field AI",
    "version": "0.1.0",
    "description": "Local prototype. All actual mutations require project review. No external scheduling writes."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/projects/{projectId}/activities": {
      "get": {
        "summary": "List project activities",
        "operationId": "get__projects_projectId_activities",
        "responses": {
          "200": {
            "description": "Successful response; reports/imports use asynchronous processing"
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Project role or CSRF check failed"
          },
          "409": {
            "description": "Stale version or idempotency conflict"
          },
          "422": {
            "description": "Business-rule validation failed"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ]
      }
    },
    "/api/projects/{projectId}/activities/{activityId}": {
      "get": {
        "summary": "Get scoped activitie",
        "operationId": "get__projects_projectId_activities_activityId",
        "responses": {
          "200": {
            "description": "Successful response; reports/imports use asynchronous processing"
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Project role or CSRF check failed"
          },
          "409": {
            "description": "Stale version or idempotency conflict"
          },
          "422": {
            "description": "Business-rule validation failed"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "fieldErrors",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}